requirement_bk.txt
# --- Core ---
apache-airflow==2.8.1

# --- Providers (保留您原本列出的) ---
apache-airflow-providers-amazon==8.16.0
apache-airflow-providers-apache-hive==8.2.1
apache-airflow-providers-apache-kafka==1.3.1
apache-airflow-providers-celery==3.5.1
apache-airflow-providers-cncf-kubernetes==7.13.0
apache-airflow-providers-common-io==1.2.0
apache-airflow-providers-common-sql==1.20.0
apache-airflow-providers-docker==3.9.1
apache-airflow-providers-elasticsearch==4.5.1
apache-airflow-providers-ftp==3.7.0
apache-airflow-providers-google==10.13.1
apache-airflow-providers-grpc==3.4.1
apache-airflow-providers-hashicorp==3.6.1
apache-airflow-providers-http==4.8.0
apache-airflow-providers-imap==3.5.0
apache-airflow-providers-microsoft-azure==8.5.1
apache-airflow-providers-mysql==5.5.1
apache-airflow-providers-odbc==4.4.0
apache-airflow-providers-openlineage==1.4.0
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-redis==3.6.0
apache-airflow-providers-sendgrid==3.4.0
apache-airflow-providers-sftp==4.8.1
apache-airflow-providers-slack==8.5.1
apache-airflow-providers-snowflake==5.2.1
apache-airflow-providers-sqlite==3.7.0
apache-airflow-providers-ssh==3.10.0

# --- User Custom Libs (保留您業務邏輯可能用到的) ---
# 只有當您的 DAG 程式碼裡有直接 import 這些庫時才保留
# 如果只是 Airflow 依賴的，請刪除
pandas==2.1.4
numpy==1.24.4
snowflake-connector-python==3.6.0
snowflake-sqlalchemy==1.5.1
mysql-connector-python==8.0.29
azure-storage-blob==12.19.0
boto3==1.33.13
paramiko==3.4.0
beautifulsoup4==4.12.2
requests==2.31.0
